Better handle imperfectly formed html text input.
This change involved switching to a different html parser library (from FSharp.Data to HtmlAgilityPack). Rather than replace the implicit dependency, the dependency on the parser library was extracted so that alternative libraries can be used instead, simply by passing an implementation of the HtmlLibAdapter interface to the relevant functions (of most interest would be Html2PlainText.tryExtractPlainText).
Changes:
- extract the dependency on the html parser library.
- implement adapters for each parser library
- verify behavioural differences between parsers (FSharp.Data vs HtmlAgilityPack).